end_consumer/requirements.txt
fastapi
uvicorn
pydantic
confluent-kafka
python-dotenv
redis
